requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
mediapipe==0.10.14
dependency-injector
python-dotenv==1.0.1
pytest==8.0.0
pandas>=2.0.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
